# api/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/chats": {
			"post": {
				"operationId": "ChatsController_createChat",
				"parameters": [
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOpenChatDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"$ref": "#/components/schemas/ChatEntity"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			},
			"get": {
				"operationId": "ChatsController_getChats",
				"parameters": [
					{
						"name": "search",
						"required": true,
						"in": "query",
						"description": "Chat name",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"description": "Limit chat",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"description": "Offset chat",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "notFavoriteChatIds",
						"required": false,
						"in": "query",
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/ChatEntity"
													}
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
		"/chats/{name}": {
			"get": {
				"operationId": "ChatsController_getChatByName",
				"parameters": [
					{
						"name": "name",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/ChatEntity"
													}
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
		"/chats/{id}/keys/keep": {
			"post": {
				"operationId": "ChatsController_keepChatKey",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/KeepKeyDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "object"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
		"/chats/{chatId}/keys/receive": {
			"post": {
				"operationId": "ChatsController_receiveKey",
				"parameters": [
					{
						"name": "chatId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "object"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
		"/chats/join": {
			"post": {
				"operationId": "ChatsController_join",
				"parameters": [
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FavoriteChatsDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/ChatEntity"
													}
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
		"/chats/leave": {
			"post": {
				"operationId": "ChatsController_leave",
				"parameters": [
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LeaveChatsDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "object"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Chats"
				]
			}
		},
...
		"/messages": {
			"post": {
				"operationId": "MessagesController_createMessage",
				"parameters": [
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMessageDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"$ref": "#/components/schemas/MessageEntity"
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				]
			},
			"get": {
				"operationId": "MessagesController_getMessages",
				"parameters": [
					{
						"name": "chatId",
						"required": true,
						"in": "query",
						"description": "Chat id",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"description": "Limit message",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"description": "Offset message",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": true
												},
												"data": {
													"type": "array",
													"items": {
														"$ref": "#/components/schemas/MessageEntity"
													}
												}
											},
											"required": [
												"success",
												"data"
											]
										},
										{
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean",
													"example": false
												},
												"data": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"data"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				]
			}
		},
		"/messages/{messageId}/files/{fileId}": {
			"get": {
				"operationId": "MessagesController_getFile",
				"parameters": [
					{
						"name": "fileId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "messageId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "x-socket-id",
						"required": true,
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Messages"
				]
			}
		}
	},
	"info": {
		"title": "PassimX",
		"description": "",
		"version": "1.1.3",
		"contact": {},
		"license": {
			"name": "MIT LICENSE",
			"url": "open platform for messaging"
		}
	},
...
	"servers": [
		{
			"url": "http://localhost:7020"
		}
	],
	"components": {
		"schemas": {
...
			"UserEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"rsaPublicKey": {
						"type": "string"
					},
					"encryptedRsaPrivateKey": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"userName",
					"rsaPublicKey",
					"encryptedRsaPrivateKey",
					"createdAt"
				]
			},
...
			"ChatKeyEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"encryptionKey": {
						"type": "string",
						"nullable": true
					},
					"received": {
						"type": "boolean"
					},
					"chat": {
						"$ref": "#/components/schemas/ChatEntity"
					},
					"user": {
						"$ref": "#/components/schemas/UserEntity"
					}
				},
				"required": [
					"id",
					"createdAt",
					"chatId",
					"userId",
					"encryptionKey",
					"received",
					"chat",
					"user"
				]
			},
			"ChatEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true,
						"maxLength": 64
					},
					"title": {
						"type": "string"
					},
					"countMessages": {
						"type": "number"
					},
					"type": {
						"type": "string"
					},
					"maxUsersOnline": {
						"type": "number"
					},
					"message": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/MessageEntity"
						}
					},
					"keys": {
						"nullable": true,
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ChatKeyEntity"
						}
					}
				},
				"required": [
					"id",
					"createdAt",
					"name",
					"title",
					"countMessages",
					"type",
					"maxUsersOnline",
					"keys"
				]
			},
			"Metadata": {
				"type": "object",
				"properties": {
					"duration": {
						"type": "number"
					},
					"loudnessData": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"transcriptionVoice": {
						"type": "object",
						"nullable": true
					}
				}
			},
			"FileEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"originalName": {
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"size": {
						"type": "number"
					},
					"mimeType": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					},
					"messageId": {
						"type": "string"
					},
					"metadata": {
						"$ref": "#/components/schemas/Metadata"
					},
					"message": {
						"$ref": "#/components/schemas/FileEntity"
					},
					"chat": {
						"$ref": "#/components/schemas/ChatEntity"
					}
				},
				"required": [
					"id",
					"createdAt",
					"key",
					"originalName",
					"chatId",
					"size",
					"mimeType",
					"fileType",
					"messageId",
					"metadata"
				]
			},
			"MessageEntity": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"format": "date-time",
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"number": {
						"type": "number"
					},
					"message": {
						"type": "string"
					},
					"parentMessageId": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"chat": {
						"$ref": "#/components/schemas/ChatEntity"
					},
					"user": {
						"nullable": true,
						"allOf": [
							{
								"$ref": "#/components/schemas/ChatEntity"
							}
						]
					},
					"parentMessage": {
						"$ref": "#/components/schemas/MessageEntity"
					},
					"files": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/FileEntity"
						}
					}
				},
				"required": [
					"id",
					"createdAt",
					"chatId",
					"number",
					"message",
					"parentMessageId",
					"type",
					"userId"
				]
			},
			"CreateOpenChatDto": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"description": "Chat name"
					}
				},
				"required": [
					"title"
				]
			},
			"KeyDto": {
				"type": "object",
				"properties": {
					"userId": {
						"type": "string",
						"description": "Публичный ключ",
						"maxLength": 64
					},
					"encryptionKey": {
						"type": "string",
						"description": "Зашифрованный ключ шифрования для чата",
						"maxLength": 8192
					}
				},
				"required": [
					"userId",
					"encryptionKey"
				]
			},
			"KeepKeyDto": {
				"type": "object",
				"properties": {
					"keys": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/KeyDto"
						}
					}
				},
				"required": [
					"keys"
				]
			},
			"ChatDto": {
				"type": "object",
				"properties": {
					"chatId": {
						"type": "string",
						"description": "Id chat"
					},
					"lastMessage": {
						"type": "number",
						"description": "Number last message"
					},
					"maxUsersOnline": {
						"type": "number",
						"description": "MaxUsersOnline message"
					}
				},
				"required": [
					"chatId",
					"lastMessage",
					"maxUsersOnline"
				]
			},
			"FavoriteChatsDto": {
				"type": "object",
				"properties": {
					"chats": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ChatDto"
						}
					}
				},
				"required": [
					"chats"
				]
			},
			"LeaveChatsDto": {
				"type": "object",
				"properties": {
					"chatIds": {
						"type": "array",
						"description": "Id chat",
						"items": {
							"type": "string",
							"format": "uuid"
						}
					}
				},
				"required": [
					"chatIds"
				]
			},
			"CreateFileDto": {
				"type": "object",
				"properties": {
					"key": {
						"type": "string"
					},
					"fileType": {
						"type": "string",
						"enum": [
							"is_voice",
							"is_media"
						]
					},
					"size": {
						"type": "number"
					},
					"mimeType": {
						"type": "string"
					},
					"originalName": {
						"type": "string"
					},
					"metadata": {
						"$ref": "#/components/schemas/Metadata"
					}
				},
				"required": [
					"key",
					"fileType",
					"size",
					"mimeType",
					"originalName"
				]
			},
			"CreateMessageDto": {
				"type": "object",
				"properties": {
					"chatId": {
						"type": "string",
						"description": "Chat id"
					},
					"message": {
						"type": "string",
						"description": "Message",
						"minLength": 1,
						"maxLength": 4096
					},
					"parentMessageId": {
						"type": "string",
						"description": "Message id"
					},
					"files": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CreateFileDto"
						}
					}
				},
				"required": [
					"chatId"
				]
			},
...
		}
	}
}